fix: Add VideoItem & HomePage missing components
The reason those components are imported but not added to the
'components' is that storybook has config enabling auto import of all
components in 'components' directory, as the components build up,
some imports are not added to 'components' in time. And the error is
not detected because storybook keeps working as usual.

Github issue #17