feat: Use string for VideoItem expanded transition
The transition between folded view and expanded view uses 'v-if' and
'v-else', but with the same tag name. To make sure each key is unique,
a Boolean is okay, but there's no need to change the key for either
clause, and the bound Boolean value 'expanded' changes, then the keys
will change, so Strings will be a better choice.

Github issue #16
Github issue #17