feat: Rewrite PlayerSlider
This rewrite changes the API. PlayerSlider now has 'value', 'max', and
'discrete' is set to 'false' by default. The step however is still '1'
for discrete sliders. This version makes sure only clicks and drags on
the visible elements active the slider. It also has a label and a
'format' function property to transform the value into desired string.

Github issue #22