feat: Implement HomePage selection toolbar buttons
The default behavior of 'favorite' toggle button always sets all
selected item 'favorite' to 'true', and it always sets selected items
'lastWatched' of those not watched to Data now. But it sets 'favorite'
to 'false' after the first toggle; and it sets selected items
'lastWatched' of those watched to 0.

Github issue #35